+++ src/leiningen/new/db.clj
@@ -11,12 +11,12 @@
     (some #{"+sqlite"} features) :sqlite))
 
 (defn db-dependencies [options]
-  [['luminus-migrations "0.2.8"]
+  [['luminus-migrations "0.2.9"]
    ['conman "0.6.2"]
    ({:postgres ['org.postgresql/postgresql "9.4.1212"]
      :mysql    ['mysql/mysql-connector-java "5.1.6"]
      :h2       ['com.h2database/h2 "1.4.192"]
-     :sqlite   ['org.xerial/sqlite-jdbc "3.8.11.2"]}
+     :sqlite   ['org.xerial/sqlite-jdbc "3.15.1"]}
      (select-db options))])
 
 (defn db-url [{:keys [sanitized] :as options} suffix]