+++ src/leiningen/new/luminus.clj
@@ -106,7 +106,7 @@
    ['ring-webjars "0.1.1"]
    ['ring/ring-defaults "0.2.1"]
    ['luminus/ring-ttl-session "0.3.1"]
-   ['mount "0.1.10"]
+   ['mount "0.1.11"]
    ['cprop "0.1.9"]
    ['org.clojure/tools.cli "0.3.5"]
    ['luminus-nrepl "0.1.4"]])